requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
stripe==7.9.0
clerk-backend-api==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
reportlab==4.0.7
PyPDF2==3.0.1
